refactor: consolidate shared instruction utilities to fix DRY violations
