+++ src/utils/instructionUtils.js
@@ -457,8 +457,83 @@ function parseLocalOperation(normalized, original) {
   return null;
 }
 
+function op(item) {
+  const insn = item && item.instruction;
+  return typeof insn === 'string' ? insn : insn && insn.op;
+}
+
+function arg(item) {
+  const insn = item && item.instruction;
+  return insn && typeof insn === 'object' ? insn.arg : null;
+}
+
+function pushValue(item) {
+  const itemOp = op(item);
+  if (itemOp === 'iconst_m1') return -1;
+  if (/^iconst_[0-5]$/.test(itemOp || '')) return Number(itemOp.slice(-1));
+  if (itemOp === 'bipush' || itemOp === 'sipush') return Number(arg(item));
+  return null;
+}
+
+function intLoadLocal(item) {
+  const itemOp = op(item);
+  if (itemOp === 'iload') return String(arg(item));
+  if (/^iload_[0-3]$/.test(itemOp || '')) return itemOp.slice(-1);
+  return null;
+}
+
+function intStoreLocal(item) {
+  const itemOp = op(item);
+  if (itemOp === 'istore') return String(arg(item));
+  if (/^istore_[0-3]$/.test(itemOp || '')) return itemOp.slice(-1);
+  return null;
+}
+
+function objectLoadLocal(item) {
+  const itemOp = op(item);
+  if (itemOp === 'aload') return String(arg(item));
+  if (/^aload_[0-3]$/.test(itemOp || '')) return itemOp.slice(-1);
+  return null;
+}
+
+function objectStoreLocal(item) {
+  const itemOp = op(item);
+  if (itemOp === 'astore') return String(arg(item));
+  if (/^astore_[0-3]$/.test(itemOp || '')) return itemOp.slice(-1);
+  return null;
+}
+
+function parseParameterDescriptors(descriptor) {
+  if (!descriptor || typeof descriptor !== 'string') return [];
+  const close = descriptor.indexOf(')');
+  if (!descriptor.startsWith('(') || close < 0) return [];
+  const params = [];
+  for (let i = 1; i < close;) {
+    const start = i;
+    while (descriptor[i] === '[') i += 1;
+    if (descriptor[i] === 'L') {
+      const semi = descriptor.indexOf(';', i);
+      if (semi < 0 || semi > close) return params;
+      params.push(descriptor.slice(start, semi + 1));
+      i = semi + 1;
+    } else {
+      params.push(descriptor.slice(start, i + 1));
+      i += 1;
+    }
+  }
+  return params;
+}
+
 module.exports = {
   getStackEffect,
   normalizeInstruction,
   parseLocalOperation,
+  op,
+  arg,
+  pushValue,
+  intLoadLocal,
+  intStoreLocal,
+  objectLoadLocal,
+  objectStoreLocal,
+  parseParameterDescriptors,
 };
